Add HTML escaping and tests
